+++ src/components/TransactionList.vue
@@ -8,7 +8,7 @@
     "
   >
     <DynamicScroller
-      :items="transactions"
+      :items="items"
       :min-item-size="45"
       :emit-update="true"
       @update="onDynamicScrollerUpdate"
@@ -20,14 +20,27 @@
         />
       </template>
 
-      <template #default="{ item, active }">
+      <template #default="{ item, active, index }">
         <DynamicScrollerItem
           :item="item"
           :active="active"
           class="p-2"
         >
+          <div
+            v-if="item.type === 'headOfGroup'"
+            class="
+              mt-8
+              mb-5
+              border-b
+              border-solid border-border-dark pb-3
+              text-center text-xl text-border-dark
+            "
+          >
+            {{ dayjs(item.transaction.date).format(groupByDateFormat) }}
+          </div>
+
           <TransactionListItem
-            :transaction="item"
+            :transaction="item.transaction"
             :account-id="account.id"
           />
         </DynamicScrollerItem>
@@ -51,6 +64,8 @@
 </template>
 
 <script lang="ts">
+  import dayjs from 'dayjs';
+
   import { defineComponent, type PropType } from 'vue';
 
   import type { Account } from '../stores/accountStore';
@@ -76,19 +91,36 @@
     },
     data(): {
       isLoadingTransactions: boolean;
+      groupBy: 'month';
+      items: Array<{
+        transaction: Transaction;
+        type: 'normal'|'headOfGroup'
+      }>;
     } {
       return {
-        isLoadingTransactions: false
+        isLoadingTransactions: false,
+        groupBy: 'month',
+        items: []
       };
     },
     computed: {
       transactions(): Array<Transaction> {
         return this.transactionStore.getByAccountId(this.account.id);
+      },
+      groupByDateFormat() {
+        return 'MM.YYYY';
       }
     },
     watch: {
       async account() {
         await this.changeAccount();
+      },
+      transactions() {
+        this.items = this.transactions.map((t, index) => ({
+          id: t.id,
+          transaction: t,
+          type: this.transactionIsHeadOfGroup(t, index) ? 'headOfGroup' : 'normal'
+        }));
       }
     },
     methods: {
@@ -112,13 +144,22 @@
         if(endIndex + 10 >= this.transactions.length) {
           await this.loadMoreTransactions();
         }
+      },
+      transactionIsHeadOfGroup(transaction: Transaction, index: number) {
+        return this.groupBy &&
+          (
+            !this.transactions[index - 1] ||
+            dayjs(transaction.date)
+              .isBefore(this.transactions[index - 1]?.date, this.groupBy)
+          );
       }
     },
     setup() {
       return {
         transactionStore: useTransactionStore(),
         transactionService: useTransactionService(),
-        splitStore: useSplitStore()
+        splitStore: useSplitStore(),
+        dayjs
       };
     },
     async mounted() {